refactor(app): updated create game flow
